Fix duplicate app folder issue
